# chars: improve character parser performance