Use the `ensure` parser constructor introduced in #139 to improve the performance and maintainability of most of the char parsers. Some are still hand-rolled because doing so eke's out a couple extra nanoseconds.

Before:

  ```
┌────────────────┬──────────┬─────────┬────────────┐
  │ Name           │ Time/Run │ mWd/Run │ Percentage │
  ├────────────────┼──────────┼─────────┼────────────┤
  │ peek_char_fail │  10.84ns │  11.00w │     72.55% │
  │ any_char       │  11.15ns │  11.00w │     74.58% │
  │ char           │  14.83ns │  17.00w │     99.21% │
  │ not_char       │  14.94ns │  17.00w │    100.00% │
  │ advance 1      │   7.98ns │  11.00w │     53.43% │
  └────────────────┴──────────┴─────────┴────────────┘
```

After:

 ```
┌────────────────┬──────────┬─────────┬────────────┐
│ Name           │ Time/Run │ mWd/Run │ Percentage │
├────────────────┼──────────┼─────────┼────────────┤
│ peek_char_fail │  10.05ns │  11.00w │     82.36% │
│ any_char       │  11.88ns │  11.00w │     97.37% │
│ char           │  12.20ns │  11.00w │    100.00% │
│ not_char       │  12.11ns │  11.00w │     99.30% │
│ advance 1      │   7.87ns │  11.00w │     64.53% │
└────────────────┴──────────┴─────────┴────────────┘
```